generalize Makefile
Rewrite Makefile to conform it to standards (or at least to
get it closer to them) and to make it reusable
Now the workflow is devided into roughly three steps: analyze
data, make tables and figures, compile the manuscript.
The source R files are renamed with a prefix that assigns them
to the first or second step. Makefiles recognizes the prefix,
so that make knows which group of file should be run first.